+++ apps/io-func-sign-support/src/app/config.ts
@@ -1,5 +1,8 @@
 import * as t from "io-ts";
-
+import {
+  ApplicationInsightsConfig,
+  getApplicationInsightsConfigFromEnvironment
+} from "@io-sign/io-sign/infra/azure/appinsights/config";
 import { pipe } from "fp-ts/lib/function";
 import * as RE from "fp-ts/lib/ReaderEither";
 
@@ -11,10 +14,12 @@ import {
   CosmosConfig,
   getCosmosConfigFromEnvironment
 } from "../infra/azure/cosmos/config";
+import { sequenceS } from "fp-ts/lib/Apply";
 
 export const Config = t.type({
   azure: t.type({
-    cosmos: CosmosConfig
+    cosmos: CosmosConfig,
+    appinsights: ApplicationInsightsConfig
   }),
   pagopa: t.type({
     tokenizer: PdvTokenizerConfig
@@ -28,12 +33,15 @@ export const getConfigFromEnvironment: RE.ReaderEither<
   Error,
   Config
 > = pipe(
-  RE.Do,
-  RE.bind("cosmos", () => getCosmosConfigFromEnvironment),
-  RE.bind("tokenizer", () => getPdvTokenizerConfigFromEnvironment),
+  sequenceS(RE.Apply)({
+    cosmos: getCosmosConfigFromEnvironment,
+    tokenizer: getPdvTokenizerConfigFromEnvironment,
+    appinsights: getApplicationInsightsConfigFromEnvironment
+  }),
   RE.map((config) => ({
     azure: {
-      cosmos: config.cosmos
+      cosmos: config.cosmos,
+      appinsights: config.appinsights
     },
     pagopa: {
       tokenizer: config.tokenizer