+++ apps/io-func-sign-events/CHANGELOG.md
@@ -1,1 +1,13 @@
 # io-func-sign-events
+
+## 1.0.2
+
+### Patch Changes
+
+- c12ce70: Fix env parsing
+
+## 1.0.1
+
+### Patch Changes
+
+- 8bcd156: Configuring setup of application insights